ai_assisted_version/pyproject.toml
[project]
name = "ai-assisted-telecom-billing"
version = "0.1.0"
description = "AI-assisted MVP for P6 telecom billing assignment"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.13.0",
    "bandit>=1.7.0",
    "bcrypt>=4.0.0",
    "fastapi>=0.115.0",
    "mypy>=1.20.0",
    "pip-audit>=2.7.0",
    "psycopg2-binary>=2.9.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-jose[cryptography]>=3.3.0",
    "redis>=5.0.0",
    "ruff>=0.15.0",
    "sqlalchemy>=2.0.0",
    "uvicorn[standard]>=0.30.0",
]

[tool.ruff]
line-length = 110

[tool.mypy]
ignore_missing_imports = true

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
